Fixed failing unit tests. Moved the thumbnail service to the thumbnail package.
